rename files to clarify they are for BabyAI only
