+++ minigrid/utils/baby_ai_bot.py
